Remove the legacy trace viewer in favor of NewTraceViewer.
Drop RunTraceView and WorkflowTraceViewer, move shared Span/Trace types into lib/trace-types, and keep timing helpers under the new viewer.
