+++ packages/web-shared/src/components/index.ts
@@ -1,3 +1,4 @@
+export type { Span, SpanEvent } from '../lib/trace-types';
 export { ErrorBoundary } from './error-boundary';
 export { EventListView } from './event-list-view';
 export type {
@@ -14,7 +15,6 @@ export {
   useHookActions,
 } from './hook-actions';
 export { TraceViewerSkeleton } from './new-trace-viewer/components/trace-viewer-skeleton';
-export { RunTraceView } from './run-trace-view';
 export { ConversationView } from './sidebar/conversation-view';
 export type {
   SelectedSpanInfo,
@@ -30,7 +30,6 @@ export type {
 } from './sidebar/use-selected-span-detail';
 export { type StreamChunk, StreamViewer } from './stream-viewer';
 export { StreamViewerSkeleton } from './stream-viewer-skeleton';
-export type { Span, SpanEvent } from './trace-viewer/types';
 export { NewTraceViewer } from './trace-viewer-new';
 export {
   DataInspector,
@@ -44,4 +43,3 @@ export { Kbd } from './ui/kbd';
 export { LoadMoreButton } from './ui/load-more-button';
 export { MenuDropdown, type MenuDropdownOption } from './ui/menu-dropdown';
 export { Spinner } from './ui/spinner';
-export { WorkflowTraceViewer } from './workflow-trace-view';
