+++ packages/web-shared/src/components/new-trace-viewer/components/event-list.tsx
@@ -1,7 +1,6 @@
 import { Circle } from 'lucide-react';
 import { useRef } from 'react';
 import { cn } from '../../../lib/cn';
-import { formatDurationPrecise } from '../../trace-viewer/util/timing';
 import { Tooltip, TooltipContent, TooltipTrigger } from '../../ui/tooltip';
 import {
   SleepIcon,
@@ -11,6 +10,7 @@ import {
 } from '../icons';
 import { isSpanDimmedBySearch, type SpanSearchResult } from '../search';
 import type { Span } from '../types';
+import { formatDurationPrecise } from '../util/timing';
 import { getSpanDurationMs, isSpanErrored } from '../utils';
 import { MiddleTruncate } from './middle-truncate/middle-truncate';
 import { ROW_HEIGHT_PX, useRowWindow } from './use-row-window';
