+++ packages/web-shared/src/components/new-trace-viewer/components/timeline.tsx
@@ -12,12 +12,9 @@ import {
   useState,
 } from 'react';
 import { cn } from '../../../lib/cn';
-import {
-  formatDurationPrecise,
-  getHighResInMs,
-} from '../../trace-viewer/util/timing';
 import { isSpanDimmedBySearch, type SpanSearchResult } from '../search';
 import type { Span } from '../types';
+import { formatDurationPrecise, getHighResInMs } from '../util/timing';
 import type {
   OffscreenMarkers,
   Segment,
