+++ packages/web-shared/src/components/new-trace-viewer/types.ts
@@ -1,2 +1,2 @@
 export type { TraceWithMeta } from '../../lib/trace-builder';
-export type { Span, SpanEvent, Trace } from '../trace-viewer/types';
+export type { Span, SpanEvent, Trace } from '../../lib/trace-types';
