+++ packages/web-shared/src/components/new-trace-viewer/util/timing.ts
