--- packages/web-shared/src/components/trace-viewer/components/zoom-icons.tsx
@@ -1,65 +1,0 @@
-import type { ReactNode } from 'react';
-
-export function IconZoomIn({ size = 16 }: { size?: number }): ReactNode {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" aria-hidden>
-      <title>Zoom In</title>
-      <path
-        d="M11 19a8 8 0 1 0 0-16 8 8 0 0 0 0 16Z"
-        stroke="currentColor"
-        strokeWidth="2"
-      />
-      <path
-        d="M11 8v6M8 11h6"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-      />
-      <path
-        d="M20 20l-3.5-3.5"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-      />
-    </svg>
-  );
-}
-
-export function IconZoomOut({ size = 16 }: { size?: number }): ReactNode {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" aria-hidden>
-      <title>Zoom Out</title>
-      <path
-        d="M11 19a8 8 0 1 0 0-16 8 8 0 0 0 0 16Z"
-        stroke="currentColor"
-        strokeWidth="2"
-      />
-      <path
-        d="M8 11h6"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-      />
-      <path
-        d="M20 20l-3.5-3.5"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-      />
-    </svg>
-  );
-}
-
-export function IconMinus({ size = 16 }: { size?: number }): ReactNode {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" aria-hidden>
-      <title>Minus</title>
-      <path
-        d="M6 12h12"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-      />
-    </svg>
-  );
-}
