--- packages/web-shared/src/components/trace-viewer/index.tsx
@@ -1,4 +1,0 @@
-export { PanelResizer } from './components/span-detail-panel';
-export { TraceViewerContextProvider, useTraceViewer } from './context';
-export * from './trace-viewer';
-export type { Trace } from './types';
