--- packages/web-shared/src/components/trace-viewer/util/constants.ts
@@ -1,8 +1,0 @@
-export const SEARCH_HEIGHT = 40;
-export const SEARCH_GAP = 8;
-export const MAP_HEIGHT = 56;
-export const TIMELINE_PADDING = 8;
-export const ROW_PADDING = 2;
-export const ROW_HEIGHT = 24;
-export const MARKER_HEIGHT = 24;
-export const MARKER_NOTCH_HEIGHT = 8;
